Move periodic tasks into ShenandoahHeap
