+++ src/hotspot/share/gc/shenandoah/shenandoahControlThread.cpp
@@ -50,31 +50,16 @@ ShenandoahControlThread::ShenandoahControlThread() :
   ConcurrentGCThread(),
   _alloc_failure_waiters_lock(Mutex::safepoint-2, "ShenandoahAllocFailureGC_lock", true),
   _gc_waiters_lock(Mutex::safepoint-2, "ShenandoahRequestedGC_lock", true),
-  _periodic_task(this),
   _requested_gc_cause(GCCause::_no_cause_specified),
   _degen_point(ShenandoahGC::_degenerated_outside_cycle),
   _allocs_seen(0) {
   set_name("Shenandoah Control Thread");
   reset_gc_id();
   create_and_start();
-  _periodic_task.enroll();
-  if (ShenandoahPacing) {
-    _periodic_pacer_notify_task.enroll();
-  }
 }
 
 ShenandoahControlThread::~ShenandoahControlThread() {
   // This is here so that super is called.
-}
-
-void ShenandoahPeriodicTask::task() {
-  _thread->handle_force_counters_update();
-  _thread->handle_counters_update();
-}
-
-void ShenandoahPeriodicPacerNotify::task() {
-  assert(ShenandoahPacing, "Should not be here otherwise");
-  ShenandoahHeap::heap()->pacer()->notify_waiters();
 }
 
 void ShenandoahControlThread::run_service() {
@@ -195,7 +180,7 @@ void ShenandoahControlThread::run_service() {
 
       // If GC was requested, we are sampling the counters even without actual triggers
       // from allocation machinery. This captures GC phases more accurately.
-      set_forced_counters_update(true);
+      heap->set_forced_counters_update(true);
 
       // If GC was requested, we better dump freeset data for performance debugging
       {
@@ -236,16 +221,16 @@ void ShenandoahControlThread::run_service() {
         // Notify Universe about new heap usage. This has implications for
         // global soft refs policy, and we better report it every time heap
         // usage goes down.
-        Universe::heap()->update_capacity_and_used_at_gc();
+        heap->update_capacity_and_used_at_gc();
 
         // Signal that we have completed a visit to all live objects.
-        Universe::heap()->record_whole_heap_examined_timestamp();
+        heap->record_whole_heap_examined_timestamp();
       }
 
       // Disable forced counters update, and update counters one more time
       // to capture the state at the end of GC session.
-      handle_force_counters_update();
-      set_forced_counters_update(false);
+      heap->handle_force_counters_update();
+      heap->set_forced_counters_update(false);
 
       // Retract forceful part of soft refs policy
       heap->soft_ref_policy()->set_should_clear_all_soft_refs(false);
@@ -581,28 +566,8 @@ void ShenandoahControlThread::notify_gc_waiters() {
   ml.notify_all();
 }
 
-void ShenandoahControlThread::handle_counters_update() {
-  if (_do_counters_update.is_set()) {
-    _do_counters_update.unset();
-    ShenandoahHeap::heap()->monitoring_support()->update_counters();
-  }
-}
-
-void ShenandoahControlThread::handle_force_counters_update() {
-  if (_force_counters_update.is_set()) {
-    _do_counters_update.unset(); // reset these too, we do update now!
-    ShenandoahHeap::heap()->monitoring_support()->update_counters();
-  }
-}
-
 void ShenandoahControlThread::notify_heap_changed() {
   // This is called from allocation path, and thus should be fast.
-
-  // Update monitoring counters when we took a new region. This amortizes the
-  // update costs on slow path.
-  if (_do_counters_update.is_unset()) {
-    _do_counters_update.set();
-  }
   // Notify that something had changed.
   if (_heap_changed.is_unset()) {
     _heap_changed.set();
@@ -612,10 +577,6 @@ void ShenandoahControlThread::notify_heap_changed() {
 void ShenandoahControlThread::pacing_notify_alloc(size_t words) {
   assert(ShenandoahPacing, "should only call when pacing is enabled");
   Atomic::add(&_allocs_seen, words, memory_order_relaxed);
-}
-
-void ShenandoahControlThread::set_forced_counters_update(bool value) {
-  _force_counters_update.set_cond(value);
 }
 
 void ShenandoahControlThread::reset_gc_id() {
